% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp.meta.R
\name{cp.meta}
\alias{cp.meta}
\title{Function to meta-analyse raw data summary statistics with external
summary statistics}
\usage{
cp.meta(cp.plink.linear, cor, chr = NULL, pos = NULL, snp = NULL,
  A1 = NULL, A2 = NULL, nonmiss = NULL, weight.pl = NULL,
  exclude.ambiguous = TRUE)
}
\arguments{
\item{cp.plink.linear}{An \code{cp.plink.linear} object}

\item{cor, chr, pos, snp, A1, A2}{see \code{\link{lassosum.pipeline}}}

\item{nonmiss}{Number of non-missing observations (can be either a scalar or a vector)}

\item{weight.pl}{An arbitrary weight between 0 and 1 given to the plink.linear correlation}

\item{exclude.ambiguous}{Should ambiguous SNPs be excluded?}
}
\value{
A \code{\link{cp.plink.linear}} object
}
\description{
Function to meta-analyse raw data summary statistics with external
summary statistics
}
\details{
This function performs a meta-analysis of the correlations 
coefficients as calculated in \code{cp.plink.linear} and some external 
correlations. If correlation coefficients are not available, these
can be converted from p-values using \code{\link{p2cor}}.
}
\note{
Either \code{n} or \code{nonmiss} or \code{weight.pl} must be specified
}
