% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp.pthresh.R
\name{cp.pthresh}
\alias{cp.pthresh}
\title{p-value thresholding with cross-prediction}
\usage{
cp.pthresh(cp.plink.linear, pvals = NULL, ...)
}
\arguments{
\item{cp.plink.linear}{A \code{\link{cp.plink.linear}} object}

\item{...}{Other parameters to pass to lassosum.pipeline()}
}
\value{
A \code{cp.lassosum} object
}
\description{
A wrapper around \code{\link{cp.lassosum}}
}
