% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgs.pthresh.R
\name{.pgs.pthresh}
\alias{.pgs.pthresh}
\title{Internal engine for polygenic scores with p-value thresholding}
\usage{
\method{.pgs}{pthresh}(weights, bfile, keep = NULL, extract = NULL,
  exclude = NULL, remove = NULL, chr = NULL, cluster = NULL,
  trace = 0)
}
\arguments{
\item{weights}{A pthresh object!!! (I use this name to trick validate.lassosum.pipeline)}

\item{bfile}{A plink bfile stem}

\item{keep}{samples to keep (see \code{\link{parseselect}})}

\item{extract}{SNPs to extract (see \code{\link{parseselect}})}

\item{exclude}{SNPs to exclude (see \code{\link{parseselect}})}

\item{remove}{samples to remove (see \code{\link{parseselect}})}

\item{chr}{a vector of chromosomes}

\item{cluster}{A \code{cluster} object from the \code{parallel} package.
For parallel processing.}
}
\value{
A matrix of Polygenic Scores
}
\description{
Internal engine for polygenic scores with p-value thresholding
}
