% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.lassosum.pipeline.R
\name{merge.lassosum.pipeline}
\alias{merge.lassosum.pipeline}
\title{merge.lassosum.pipeline for lassosum.pipeline objects produced by pthresh.pipeline}
\usage{
\method{merge}{lassosum.pipeline}(..., pthresh = F)
}
\description{
merge.lassosum.pipeline for lassosum.pipeline objects produced by pthresh.pipeline
}
