% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiBed4}
\alias{multiBed4}
\title{Mutiply a bed file with a vector with p-value thresholds bins}
\usage{
multiBed4(fileName, N, P, weights, pbin, nbin, col_skip_pos, col_skip,
  keepbytes, keepoffset, trace)
}
\description{
Mutiply a bed file with a vector with p-value thresholds bins
}
