% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.linear.R, R/plink.parse.R
\name{plink.linear}
\alias{plink.linear}
\alias{plink.parse}
\title{Obtain standardized coefficients from linear regression in PLINK}
\usage{
plink.linear(bfile, pheno, out = tempfile("lassosum.out"), keep = NULL,
  remove = NULL, extract = NULL, exclude = NULL, chr = NULL,
  covar = NULL, plink.cmd = NULL, extension = NULL,
  ignore.stdout = trace <= 0, trace = 0,
  plink = getOption("lassosum.plink"), plink2 = grepl("plink2$",
  plink), ..., parsed = NULL)

plink.parse(parsed, out, plink, pheno = NULL, covar = NULL, ...,
  details = FALSE)
}
\arguments{
\item{pheno}{Phenotype. see details.}

\item{out}{The \code{--out} option in plink}

\item{keep, remove, extract, exclude, chr}{see parseselect()}

\item{covar}{Covariates. see details}

\item{plink.cmd}{The command passed to plink}

\item{extension}{The extension for the output file generated from the plink command}

\item{ignore.stdout}{Ignore stdout. see \code{\link[base]{base::system}}}

\item{plink}{plink executive}

\item{plink2}{Whether we are using plink2}

\item{...}{Other parameters to pass to plink. See details}

\item{plink}{plink executive}

\item{...}{Other parameters to pass to plink. See details}
}
\description{
Obtain standardized coefficients from linear regression in PLINK

parse options parsed to PLINK
}
\details{
\code{pheno} and \code{covar} can take one of three formats:
         \itemize{
         \item a vector/matrix with length/number of rows equal the 
         number of included samples, or
         \item a data.frame with the 
         first two columns labelled "FID" and "IID" and the other columns
         giving the phenotype/covariates, or
         \item the name of a file in a format 
         understood by plink's \code{--pheno} or \code{--covar} options
         } 
         Only one column can be given for \code{pheno}, although covar can take 
         more than one columns. 
         
         \code{...}: Other options to plink can be given by, e.g., 
         \code{keep.allele.order=T} for \code{--keep-allele-order}, 
         \code{maf=0.01} for \code{--maf 0.01}, etc.
}
