% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.pthresh.R
\name{[.pthresh}
\alias{[.pthresh}
\title{Subsetting a matrix of p-value threshold betas}
\usage{
\method{[}{pthresh}(obj, row = 1:length(obj$beta), col = 1:obj$nbin,
  pthresh = FALSE)
}
\description{
Subsetting a matrix of p-value threshold betas
}
