% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.lassosum.pipeline.R
\name{subset.lassosum.pipeline}
\alias{subset.lassosum.pipeline}
\title{subsetting a lassosum.pipeline object, by s and lambda. Alternatively, 
provide a validate.lassosum object.}
\usage{
\method{subset}{lassosum.pipeline}(lassosum.pipeline, validate.lassosum, s,
  lambda)
}
\description{
subsetting a lassosum.pipeline object, by s and lambda. Alternatively, 
provide a validate.lassosum object.
}
